#18 Reorganize code to server module
